% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{addDDCoords}
\alias{addDDCoords}
\title{addDDCoords}
\usage{
addDDCoords(GSINF_ = NULL)
}
\arguments{
\item{GSINF_}{original GSINF object}
}
\description{
addDDCoords
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
