% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{disentangleGSDET}
\alias{disentangleGSDET}
\title{disentangleGSDET}
\usage{
disentangleGSDET(GSDET_ = NULL, GSCAT_ = NULL, GSINF_ = NULL)
}
\arguments{
\item{GSDET_}{directory}

\item{GSCAT_}{directory}

\item{GSINF_}{directory}
}
\description{
This function creates
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
