% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor.R
\name{extractor}
\alias{extractor}
\title{extractor}
\usage{
extractor(tblNames = NULL, schema = NULL, uid = NULL, pw = NULL, dsn = "PTRAN")
}
\description{
This function
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
