% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{fixHerringLengths}
\alias{fixHerringLengths}
\title{fixHerringLengths}
\usage{
fixHerringLengths(GSDET_ = NULL)
}
\arguments{
\item{GSDET_}{directory}
}
\description{
This function creates
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
