% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSAUX}
\alias{GSAUX}
\title{GSAUX Metadata}
\format{
data.frame
\describe{
\item{AUX}{A single digit numeric code to identify any auxiliary gear being used.}
\item{AUXDESC}{Description of auxiliary gear.}
}
}
\usage{
GSAUX
}
\description{
This is auxiliary gear code table.
}
\keyword{datasets}
