% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSCURNT}
\alias{GSCURNT}
\title{GSCURNT Metadata}
\format{
data.frame
\describe{
\item{CURNT}{A single digit code to identify the direction of the tide in relation to the vessel}
\item{CURNTDESC}{The description of CURNT (e.g. "starboard")}
}
}
\usage{
GSCURNT
}
\description{
This is a lookup table for describing the direction of the tide relative to the vessel.
}
\keyword{datasets}
