% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSDET}
\alias{GSDET}
\title{GSDET Metadata}
\format{
data.frame
\describe{
\item{MISSION}{A 10 character field to identify the survey.  The first 3 character identifies the vessel (eg. NED), the next 4 characters identifies the year (eg. 2000) and the last 3 characters identifies the survey number within the year. (eg. 001). Example NED2000001.}
\item{SETNO}{Each tow is assigned a numeric set identifier starting with 1 for the first tow and then assigned consecutively.}
\item{SPEC}{Maritimes species code.}
\item{FSHNO}{Unique number assigned within species and mission.}
\item{FLEN}{xxx}
\item{FSEX}{Sex of the specimen.}
\item{FMAT}{Code for the observed maturity stage.}
\item{FWT}{The round weight in grams of the specimen.}
\item{AGMAT}{Code for the age material collected.}
\item{NANN}{Number of Annuli.}
\item{EDGE}{Edge Type}
\item{CHKMRK}{Check Mark is made up of 3 single character fields.}
\item{AGE}{Age of specimen.}
\item{AGER}{Ager code}
\item{CLEN}{Count at Length.}
\item{REMARKS}{Any comments pertaining to the specimen.}
\item{SIZE_CLASS}{Size Class.}
\item{SPECIMEN_ID}{Specimen Identifier}
}
}
\usage{
GSDET
}
\description{
This table describes the various measurements take for individual specimens (e.g. length, weight, etc).
}
\keyword{datasets}
