% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSFORCE}
\alias{GSFORCE}
\title{GSFORCE Metadata}
\format{
data.frame
\describe{
\item{FORCE}{A single digit code to identify wind force on the Beaufort scale}
\item{FORCEDESC}{The description of FORCE (e.g. "1 - 3 Knots")}
}
}
\usage{
GSFORCE
}
\description{
This is a lookup table for describing the wind.
}
\keyword{datasets}
