% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSGEAR}
\alias{GSGEAR}
\title{GSGEAR Metadata}
\format{
data.frame
\describe{
\item{GEAR}{A single digit numeric code to identify the gear being used.}
\item{GEARDESC}{The description of GEAR (e.g. "Western IIA trawl")}
}
}
\usage{
GSGEAR
}
\description{
This is a lookup table for describing the gear
}
\keyword{datasets}
