% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSHOWOBT}
\alias{GSHOWOBT}
\title{GSHOWOBT Metadata}
\format{
data\.frame
\describe{
\item{HOWOBT}{A single digit numeric code describing how the ship's position was determined.}
\item{HOWOBTDESC}{The description of HOWOBT (e.g. "GPS")}
}
}
\usage{
GSHOWOBT
}
\description{
This is a lookup table for describing how the vessel's positional information was obtained.
}
\keyword{datasets}
