% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSINF}
\alias{GSINF}
\title{GSINF Metadata}
\format{
data.frame
\describe{
\item{MISSION}{A 10 character field to identify the survey. The first 3 character identifies the vessel (eg. NED), the next 4 characters identifies the year (eg. 2000) and the last 3 characters identifies the survey number within the year. (eg. 001). Example NED2000001.}
\item{SETNO}{Each tow is assigned a numeric set identifier starting with 1 for the first tow and then assigned consecutively.}
\item{SDATE}{The date at the beginning of the tow.}
\item{TIME}{The local time at the beginning of the tow.}
\item{STRAT}{A numeric code to identify stratum.}
\item{AREA}{A numeric code to identify the NAFO area.}
\item{DUR}{Duration.}
\item{DIST}{The actual tow distance in nautical miles.}
\item{SPEED}{The average speed of the vessel over bottom, based on GPS, to the nearest tenth of a nautical mile.}
\item{HOWD}{A single digit code to identify the source of information for the distance.}
\item{HOWS}{A single digit code to identify the source of information for the speed.}
\item{WIND}{Wind Direction. In true degrees, between 0 and 360.}
\item{FORCE}{A single digit code to identify wind force on the Beaufort scale}
\item{CURNT}{A single digit code to identify the direction of the tide in relation to the vessel}
\item{TYPE}{A numeric code to identify the experiment type.}
\item{GEAR}{A single digit numeric code to identify the gear being used.}
\item{AUX}{A single digit numeric code to identify any auxiliary gear being used.}
\item{REMARKS}{Any comments pertaining to set.}
\item{SURFACE_TEMPERATURE}{Surface Temperature in celsius}
\item{BOTTOM_TEMPERATURE}{Bottom temperature in celsius }
\item{BOTTOM_SALINITY}{Bottom salinity measured in pounds per square unit. }
\item{STATION}{xxx}
\item{SLAT_DD}{The latitude at the start of a tow (decimal degrees)}
\item{SLONG_DD}{The longitude at the start of a tow (decimal degrees)}
\item{ELAT_DD}{The latitude at the end of a tow (decimal degrees)}
\item{ELONG_DD}{The longitude at the end of a tow (decimal degrees)}
\item{DMIN_M}{The minimum(shallowest) bottom depth, in meters observed during a tow.}
\item{DMAX_M}{The maximum(deepest) bottom depth, in meters observed during a tow.}
\item{DEPTH_M}{The average of start and end depth}
\item{START_DEPTH_M}{The bottom depth at start of a set, in meters}
\item{END_DEPTH_M}{The bottom depth at end of a set, in meters}
}
}
\usage{
GSINF
}
\description{
This table describes information about the sets (e.g. when and where they occurred)
}
\keyword{datasets}
