% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSMATURITY}
\alias{GSMATURITY}
\title{GSMATURITY Metadata}
\format{
data.frame
\describe{
\item{CODE}{Maturity Code}
\item{DESCRIPTION}{Description of maturity stage}
}
}
\usage{
GSMATURITY
}
\description{
This is the code table for the maturity codes.
}
\keyword{datasets}
