% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSMISSIONS}
\alias{GSMISSIONS}
\title{GSMISSIONS Metadata}
\format{
data.frame
\describe{
\item{MISSION}{A 10 character field to identify the survey.  The first 3 character identifies the vessel (eg. NED), the next 4 characters identifies the year (eg. 2000) and the last 3 characters identifies the survey number within the year. (eg. 001). Example NED2000001.}
\item{VESEL}{A 1 digit code used to identify the vessel used in the survey}
\item{CRUNO}{A 3 digit code identifying the survey number within the year. (eg. 001)}
\item{YEAR}{The year of the survey.}
\item{SEASON}{The season of the survey.}
}
}
\usage{
GSMISSIONS
}
\description{
This table describes all of the different survey cruises (e.g. when and where they went, etc)
}
\keyword{datasets}
