% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSEX}
\alias{GSSEX}
\title{GSSEX Metadata}
\format{
data.frame
\describe{
\item{CODE}{Sex Code}
\item{DESCRIPTION}{Description of sex code}
}
}
\usage{
GSSEX
}
\description{
This is the code table for the sex codes.
}
\keyword{datasets}
