% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSPEC}
\alias{GSSPEC}
\title{GSSPEC Metadata}
\format{
data.frame
\describe{
\item{SPEC}{Maritimes species code.}
\item{CNAME}{The common name of the species.}
\item{DMIN}{Minimum depth for this species}
\item{DMAX}{Maximum depth for this species}
\item{VSEX1}{xxx}
\item{VSEX2}{xxx}
\item{VSEX3}{xxx}
\item{LGRP}{length grouping}
\item{LMIN}{Minimum length}
\item{LMAX}{Maximum length}
\item{MATREQ}{Aging materials required}
\item{SAMPREQ}{Samples required}
\item{MLEN}{xxx}
\item{LENWGTA}{Length weight regression coefficient a}
\item{LENWGTB}{Length weight regression coefficient b}
\item{LFSEXED}{Length frequency counts by sex (Y/N)}
\item{MAXWGT}{Maximum weight for species}
}
}
\usage{
GSSPEC
}
\description{
This table appears to describe quality control and data processing parameters by species
}
\keyword{datasets}
