% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSPECIES_TAX}
\alias{GSSPECIES_TAX}
\title{GSSPECIES_TAX}
\format{
data.frame
\describe{
\item{APHIAID}{WoRMS Aphia ID code}
\item{KINGDOM}{KINGDOM}
\item{PHYLUM}{PHYLUM}
\item{CLASS}{CLASS}
\item{ORDER}{ORDER}
\item{FAMILY}{FAMILY}
\item{GENUS}{GENUS}
\item{SPECIES}{SPECIES}
}
}
\usage{
GSSPECIES_TAX
}
\description{
This is a table of the taxonomic hierarchies associated with the APHIAIDs found in GSSPECIES_20220624 (from WoRMS)
}
\keyword{datasets}
