% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSTRATA}
\alias{GSSTRATA}
\title{GSSTRATA Metadata}
\format{
data.frame
\describe{
\item{STRATA}{survey stratum code}
\item{MAJOR}{xxx}
\item{MINOR}{xxx}
\item{LATITUDE}{xxx}
\item{LONGITUDE}{xxx}
}
}
\usage{
GSSTRATA
}
\description{
This appears to be a table that could be used to plot the strata.
}
\keyword{datasets}
