% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSTRATUM}
\alias{GSSTRATUM}
\title{GSSTRATUM Metadata}
\format{
data.frame
\describe{
\item{STRAT}{survey stratum code}
\item{DMIN}{Stratum minimum Depth}
\item{DMAX}{Stratum maximum Depth}
\item{AREA}{Area of stratum (square nautical miles)}
\item{DEPTH}{Stratum nominal Depth}
\item{NAME}{stratum name}
}
}
\usage{
GSSTRATUM
}
\description{
This is the depth range and area by survey stratum
}
\keyword{datasets}
