% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSVESSEL}
\alias{GSVESSEL}
\title{GSVESSEL Metadata}
\format{
data.frame
\describe{
\item{VESEL}{A 1 digit code used to identify the vessel used in the survey}
\item{VNAME}{Name of the vessel}
\item{BDATE}{xxx}
\item{EDATE}{xxx}
\item{GEAR1}{xxx}
\item{GEAR2}{xxx}
\item{GEAR3}{xxx}
\item{GEAR4}{xxx}
\item{GEAR5}{xxx}
}
}
\usage{
GSVESSEL
}
\description{
This is the...
}
\keyword{datasets}
