% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSWARPOUT}
\alias{GSWARPOUT}
\title{GSWARPOUT Metadata}
\format{
data.frame
\describe{
\item{MISSION}{A 10 character field to identify the survey.  The first 3 character identifies the vessel (eg. NED), the next 4 characters identifies the year (eg. 2000) and the last 3 characters identifies the survey number within the year. (eg. 001). Example NED2000001.}
\item{SETNO}{Each tow is assigned a numeric set identifier starting with 1 for the first tow and then assigned consecutively.}
\item{WARPOUT}{Warp, in meters}
}
}
\usage{
GSWARPOUT
}
\description{
This is the table of the WARP associated with each fishing event.
}
\keyword{datasets}
