% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSXTYPE}
\alias{GSXTYPE}
\title{GSXTYPE Metadata}
\format{
data.frame
\describe{
\item{XTYPE}{Numeric code for the experiment type}
\item{XTYPEDESC}{Description of the experiment type code }
}
}
\usage{
GSXTYPE
}
\description{
This is the code table for the (GSINF) TYPE codes.
}
\keyword{datasets}
