% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTbls.R
\name{listTbls}
\alias{listTbls}
\title{listTbls}
\usage{
listTbls()
}
\description{
This function returns the names of all of the tables available in the package
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
