% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listToFolder.R
\name{listToFolder}
\alias{listToFolder}
\title{listToFolder}
\usage{
listToFolder(dfList = NULL, dataPath = NULL)
}
\description{
This function
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
