% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{makeGSSPECIES_NEW}
\alias{makeGSSPECIES_NEW}
\title{makeGSSPECIES_NEW}
\usage{
makeGSSPECIES_NEW(GSSPECIES_ANDES_ = NULL, GSSPEC_ = NULL)
}
\arguments{
\item{GSSPECIES_ANDES_}{directory}

\item{GSSPEC_}{directory}
}
\description{
This function creates
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
