% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data.frame}
\name{maritimesBathy}
\alias{maritimesBathy}
\title{maritimesBathy Metadata}
\format{
a data.frame storing coordinates and depths in DD
\describe{
\item{x}{Longitudes (dd)}
\item{y}{Latitudes (dd)}
\item{z}{Depth (metrs)}
}
}
\usage{
maritimesBathy
}
\description{
This is a dataframe of the bathymetry of the Maritimes area, extracted via \code{getNOAA.bathy()} and \code{fortify.bathy()} functions of the \code{marmap} package on Oct 28, 2022.
}
\keyword{datasets}
