% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{maritimesCoast_sf}
\alias{maritimesCoast_sf}
\title{maritimesCoast_sf Metadata}
\format{
a data.frame storing coordinates in DD
\describe{
\item{long}{Longitudes (dd)}
\item{lat}{Latitudes (dd)}
\item{group}{This is a grouping variable for each polygon. A region or subregion might have multiple polygons, for example, if it includes island}
\item{order}{The order to connect each point within a group.}
\item{region}{Roughly, the names of countries, although some other objects are present (such as 
some lakes).}
\item{subregion}{The names of subregions within a region, which can contain multiple groups. For 
example, the Nova Scotia subregion includes many islands, each with its own group.}
}
}
\usage{
maritimesCoast_sf
}
\description{
This is a the general area of the Maritimes coast, extracted via the 
\code{mapdata} package on Apr 30, 2024. Field descriptions taken from 
https://r-graphics.org/recipe-miscgraph-map.
}
\keyword{datasets}
