% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{rmSizeClasses}
\alias{rmSizeClasses}
\title{rmSizeClasses}
\usage{
rmSizeClasses(GSCAT_ = NULL)
}
\arguments{
\item{GSCAT_}{original GSCAT object}
}
\description{
rmSizeClasses
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
