% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{RVSurveyDataTables}
\alias{RVSurveyDataTables}
\title{RVSurveyDataTables Metadata}
\format{
A character vector where each element is the name of a data object in the RVSurveyData package.
}
\usage{
RVSurveyDataTables
}
\description{
This is a character vector of the names of data objects available within the 
RVSurveyData package.
}
\keyword{datasets}
