% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{strataMar_sf}
\alias{strataMar_sf}
\title{strataMar_sf Metadata}
\format{
an sf object with a CRS of 4326 (i.e. WGS84)
\describe{
\item{STRATA_ID}{This is the identifier used for each strata}
\item{AREA_KM}{This is the area for each strata in km^2}
}
}
\usage{
strataMar_sf
}
\description{
This is an sf object of the strata used by DFO Maritimes.
}
\keyword{datasets}
