% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackageData.R
\name{updatePackageData}
\alias{updatePackageData}
\title{updatePackageData}
\usage{
updatePackageData(tblList = NULL)
}
\arguments{
\item{tblList}{default is \code{NULL}. This is a list full of named dataframes that will be 
added to the package.}
}
\description{
This function
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
