% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackageSpatialData.R
\name{updatePackageSpatialData}
\alias{updatePackageSpatialData}
\title{updatePackageSpatialData}
\usage{
updatePackageSpatialData()
}
\description{
This function generates the sf files included in RVSurveyData
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
  #' created generously large area (based on the strata) 
  #' for which to extract coastline and bathy
  maritimesLand <- ggplot2::map_data("world2Hires",region = c('Canada', 'USA', 'France'), wrap = c(-180,180))
  maritimesBathy <- marmap::fortify.bathy(marmap::getNOAA.bathy(lon1 = limits[1], 
                                                          lon2 = limits[2], 
                                                          lat1 = limits[3], 
                                                          lat2 = limits[4], 
                                                          resolution = 1))
}
