% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSPECIES}
\alias{GSSPECIES}
\title{GSSPECIES}
\format{
data.frame
\describe{
\item{CODE}{The numeric code used to identify the species/group.}
\item{UUID}{A unique identifier used to identify the species/group.}
\item{APHIAID}{WoRMS Aphia ID code.}
\item{COMM}{The common name of the species.}
\item{SPEC}{The scientific name of the species/group.}
\item{RANK}{The taxonomic level for the value of SCIENTIFICNAME}
\item{KINGDOM}{KINGDOM (for APHIAID from WoRMS)}
\item{PHYLUM}{PHYLUM (for APHIAID from WoRMS)}
\item{CLASS}{CLASS (for APHIAID from WoRMS)}
\item{ORDER}{ORDER (for APHIAID from WoRMS)}
\item{FAMILY}{FAMILY (for APHIAID from WoRMS)}
\item{GENUS}{GENUS (for APHIAID from WoRMS)}
\item{COMMENTS}{Additional information.}
\item{LGRP}{length grouping}
\item{LFSEXED}{Length frequency counts by sex (Y/N)}
}
}
\usage{
GSSPECIES
}
\description{
This is an updated version of the Maritimes Species Code table.
}
\keyword{datasets}
