% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GSSPECIES}
\alias{GSSPECIES}
\title{GSSPECIES Metadata}
\format{
data.frame
\describe{
\item{SPEC}{The scientific name of the species.}
\item{COMM}{The common name of the species.}
\item{CODE}{The numeric code used to identify the species.}
\item{NMFS}{The numeric code used by NMFS to identify the species.}
\item{ENTR}{The date the code was added.}
\item{AUTHORITY}{xxx}
\item{TSN}{The ITIS Taxonomic Serial Number.}
\item{COMMENTS}{xxx}
}
}
\usage{
GSSPECIES
}
\description{
This is the Maritimes Species Code table.
}
\keyword{datasets}
