% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateRawData.R
\name{makeGSSPECIES_NEW}
\alias{makeGSSPECIES_NEW}
\title{makeGSSPECIES_NEW}
\usage{
makeGSSPECIES_NEW(GSSPECIES_ANDES_ = NULL, GSSPEC_ = NULL)
}
\arguments{
\item{GSSPECIES_ANDES_}{original GSSPECIES_ANDES object}

\item{GSSPEC_}{original GSSPEC object}
}
\description{
This function merges GSSPEC and GSSPECIES_ANDES into a single object
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
