% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{nafo_sf}
\alias{nafo_sf}
\title{nafo_sf Metadata}
\format{
an sf object with a CRS of 4326 (i.e. WGS84)
\describe{
\item{NAFO}{This is the identifier for each NAFO subunit}
}
}
\usage{
nafo_sf
}
\description{
This is an sf object of the NAFO subunits used by DFO Maritimes.
}
\keyword{datasets}
