% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePackageData.R
\name{updatePackageData}
\alias{updatePackageData}
\title{updatePackageData}
\usage{
updatePackageData(dfList = NULL)
}
\description{
This function
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
